refactor: renamed model product => ProductModel
