+++ components/Product.tsx
@@ -1,4 +1,4 @@
-import { ProductModel } from '@/models/product';
+import { ProductModel } from '@/models/ProductModel';
 import {
     PlusSmallIcon,
     EllipsisHorizontalIcon,
